crud-flask/requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1